Fix count_ShiftLeft

+++ Core/Inc/GSM.h
@@ -98,7 +98,7 @@ Queue_GSM result_addr_queue;
 int is_disconnect = 0;
 int is_using_flash = 0;
 int mode = 0;
-int count_shiftleft = 0;
+uint8_t count_shiftleft = 0;
 
 int is_flash_overflow = 0;
 volatile uint32_t start_addr_disconnect = 0;
